+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/providers/SpecificationJsonSchemaFileProvider.kt
@@ -17,7 +17,7 @@
 
 package com.explyt.spring.web.providers
 
-import com.explyt.spring.web.model.OpenApiSpecificationDetection
+import com.explyt.spring.web.model.OpenApiSpecificationFinder
 import com.explyt.spring.web.model.OpenApiSpecificationType
 import com.explyt.spring.web.service.OpenApiSpecificationManager
 import com.intellij.openapi.application.runReadAction
@@ -28,43 +28,38 @@ import com.jetbrains.jsonSchema.extension.JsonSchemaFileProvider
 import com.jetbrains.jsonSchema.extension.SchemaType
 
 class SpecificationJsonSchemaFileProvider(
+    private val project: Project,
     private val specificationType: OpenApiSpecificationType,
-    private val remoteSchemaUrl: String,
-    private val visibleName: String,
-    private val project: Project
+    private val name: String,
+    private val remoteSource: String
 ) : JsonSchemaFileProvider {
 
     override fun isAvailable(file: VirtualFile): Boolean {
-        val detectedSpecificationType: OpenApiSpecificationType = runReadAction<Any> label@{
-            if (!file.isValid) {
-                return@label OpenApiSpecificationType.UNKNOWN.INSTANCE
-            } else {
-                val psiFile = PsiManager.getInstance(project).findFile(file)
-                if (psiFile == null) {
-                    return@label OpenApiSpecificationType.UNKNOWN.INSTANCE
-                } else {
-                    return@label OpenApiSpecificationDetection.detectPrimarySpecificationType(file, psiFile)
-                }
-            }
-        } as OpenApiSpecificationType
-        return detectedSpecificationType == specificationType
-    }
+        val specificationTypeFound = runReadAction {
+            if (!file.isValid) return@runReadAction OpenApiSpecificationType.OpenApiUndefined
 
-    override fun getName(): String {
-        return visibleName
+            val psiFile = PsiManager.getInstance(project).findFile(file)
+                ?: return@runReadAction OpenApiSpecificationType.OpenApiUndefined
+            return@runReadAction OpenApiSpecificationFinder.findSpecificationType(file, psiFile)
+        }
+
+        return specificationTypeFound == specificationType
     }
 
     override fun getSchemaFile(): VirtualFile? {
-        val schemaStorage = project.getService(OpenApiSpecificationManager::class.java)
+        return project.getService(OpenApiSpecificationManager::class.java)?.getSchemaByFile(specificationType)?.file
+    }
 
-        return schemaStorage?.getSchemaFor(specificationType)?.first
+    override fun getName(): String {
+        return name
+    }
+
+    override fun getRemoteSource(): String {
+        return remoteSource
     }
 
     override fun getSchemaType(): SchemaType {
         return SchemaType.embeddedSchema
     }
 
-    override fun getRemoteSource(): String {
-        return remoteSchemaUrl
-    }
 }